requirements.txt
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
asttokens==3.0.1
bitarray==3.8.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.1.0
debugpy==1.8.17
decorator==5.2.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.1
fonttools==4.61.0
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
parsimonious==0.10.0
parso==0.8.5
pdfkit==1.0.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py_builder_signing_sdk==0.0.2
py_clob_client==0.28.0
py_order_utils==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tenacity==9.1.2
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==14.2
